libcusmm/autotune: merge cusmm_dnt and cusmm_common
