+++ src/acc/libsmm_acc/libcusmm/kernels/cusmm_dnt.py
@@ -1,6 +1,26 @@
 # -*- coding: utf-8 -*-
 
+#===============================================================================
+# Helper variables/functions used in all kernels
 
+# Number of parameters in one stack entry
+stack_size = 16005
+npar = 3
+
+# Sizes in bytes
+sizeof_int = 4
+sizeof_double = 8
+
+
+# utility functions
+def round_up_to_multiple(x, step):
+    if x % step == 0:
+        return x
+    else:
+        return x + step - x % step
+
+
+#===============================================================================
 class Kernel:
 
     def __repr__(self):
