+++ src/acc/libsmm_acc/libcusmm/kernels/cusmm_dnt_medium.py
@@ -1,9 +1,8 @@
 # -*- coding: utf-8 -*-
-from kernels import cusmm_dnt
-import cusmm_common as cu
+from kernels import cusmm_dnt as cu
 
 
-class Kernel_dnt_medium(cusmm_dnt.Kernel):
+class Kernel_dnt_medium(cu.Kernel):
 
     algorithm = 'medium'
     launch_parameters = ['m', 'n', 'k', 'tile_m', 'tile_n', 'threads', 'grouping', 'minblocks']
